requirements.txt
spotipy
flask
python-dotenv
config
pytest
pytest-mock
pytest-logging